Serial Monitor Record - Afternoon with rain
Serial Monitor Record showing current and historical weather reports as retrieved from WU. Due to the cold and rainy weather from the day before the %Watering is still set to 0. After it starts to rain again in the afternoon a 3 hours rain delay is sent to OpenSprinkler. This rain delay was  extended by one more hour as it rains for nearly two hours.