Serial Monitor Record - Weather objects 2 x 31st March
Serial Monitor Record showing current and historical weather reports as retrieved from WU. This record shows the case with two 31st March weather objects. It is currently not clear if there are 7 or 8 weather objects available?  Yesterdays's weather information will be found in structure historyData[7]. In other cases it was in structure historyData[6]. This needs to be further investigated! However, the program handles cases with 8 and less than 8 weather objects now!