Update OS API Change Options Water level Command
Update OS API Change Options Water level Command to avoid the deactivation of the NTP, DHCP and logging features. The OS API provides no single command to change the water level / percentage. According to the OpenSprinkler Firmware 2.1.8 Document at least the parameters o2, o3 and o36 have to be included in the command to keep NTP, DHCP usage and logging active. If the items Sensor 1 option, Ignore Password or Special auto refresh are used the corresponding parameters o22, o25 and o48 need to be also included.